[dx] Bundle with esbuild to allow internal absolute imports
Allows imports within the library like `import /tanaris/strings.js` to
work for library clients.

Without this change, this syntax still works within the repo via
`compilerOptions.paths` in `tsconfig.json` and the `moduleNameMapper` in
`jest.config.cjs`, but these configs seem to be meaningless to library
clients.

So, now we use esbuild to bundle our library and resolve internal
imports in a way that library clients will understand.